ADD `Exercise` exception handling

+++ src/main/java/com/guiprojects/academy/services/ExerciseTypeService.java
@@ -7,6 +7,7 @@ import org.springframework.stereotype.Service;
 
 import com.guiprojects.academy.entities.ExerciseType;
 import com.guiprojects.academy.repositories.ExerciseTypeRepository;
+import com.guiprojects.academy.services.exceptions.ResourceNotFoundException;
 
 @Service
 public class ExerciseTypeService {
@@ -16,7 +17,7 @@ public class ExerciseTypeService {
 	
 	public ExerciseType findById (Long id) {
 		Optional<ExerciseType> obj = exerciseTypeRepository.findById(id);
-		return obj.get();
+		return obj.orElseThrow(() -> new ResourceNotFoundException(id));
 	}
 	
 	public ExerciseType insert (ExerciseType exerciseType) {
@@ -24,7 +25,7 @@ public class ExerciseTypeService {
 	}
 	
 	public ExerciseType update(Long id, ExerciseType objWithNewParameters) {
-		ExerciseType exerciseTypeToUpdate = exerciseTypeRepository.findById(id).get();
+		ExerciseType exerciseTypeToUpdate = exerciseTypeRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException(id));
 		
 		if(objWithNewParameters.getName() != null) exerciseTypeToUpdate.setName(objWithNewParameters.getName());	
 		if(objWithNewParameters.getMuscleGroup() != null) exerciseTypeToUpdate.setMuscleGroup(objWithNewParameters.getMuscleGroup());	
@@ -33,6 +34,8 @@ public class ExerciseTypeService {
 	}
 	
 	public void delete(Long id) {
-		exerciseTypeRepository.deleteById(id);
+		if(exerciseTypeRepository.existsById(id)) {
+			exerciseTypeRepository.deleteById(id);			
+		} else throw new ResourceNotFoundException(id);		
 	}
 }
